% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_the_list.R
\name{arrange_a_cleaned_list}
\alias{arrange_a_cleaned_list}
\title{Arrange a cleaned list into a list of arranged dataframes}
\usage{
arrange_a_cleaned_list(clean_list)
}
\arguments{
\item{clean_list}{a list of cleaned episodes' scripts}
}
\value{
a list of dataframes
}
\description{
Arrange a cleaned list into a list of arranged dataframes
}

