% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_episode.R
\name{arrange_episode}
\alias{arrange_episode}
\title{Arrange a script into a dataframe}
\usage{
arrange_episode(ep)
}
\arguments{
\item{ep}{a list. The list returned from \code{\link{clean_the_episode}} of a single episode.}
}
\value{
a dataframe containing arranged episode script
}
\description{
A function to arrange a cleaned script (the list that gets back from \code{\link{clean_the_episode}}).
Every row in the dataframe is a different person speaking.
}

