% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_episode.R
\name{clean_episode}
\alias{clean_episode}
\title{Clean the raw data of a script}
\usage{
clean_episode(ep)
}
\arguments{
\item{ep}{The raw data of a single script}
}
\value{
a string that is a cleaned script
}
\description{
The function takes a raw data of a script (html) and cleans it that only the script is left.
}

