% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stats.R
\name{count_episodes_appear}
\alias{count_episodes_appear}
\title{Counts number of episode each character appears}
\usage{
count_episodes_appear(all_scripts, season = "all")
}
\arguments{
\item{all_scripts}{a dataframe with all the scripts data arranged}

\item{season}{a number indicates the season. for all season can omit or "all"}
}
\value{
a frequency table
}
\description{
A function that counts how many episodes a character appears
}

