% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stats.R
\name{count_list_of_words}
\alias{count_list_of_words}
\title{Count frequency of a given list of words}
\usage{
count_list_of_words(all_scripts, word_vec, name = "all", season = "all",
  episode = "all")
}
\arguments{
\item{all_scripts}{dataframe with all the scripts arranged}

\item{word_vec}{a character vector where each item is a word to count}
}
\value{
a frequency table
}
\description{
A function that counts the frequency of a given list of words in the show
}

