% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stats.R
\name{count_num_words}
\alias{count_num_words}
\title{Count who speakes most by number of words}
\usage{
count_num_words(all_scripts)
}
\arguments{
\item{all_scripts}{a dataframe with all the scripts data arranged}
}
\value{
a frequency table
}
\description{
A function that counts who speaks the most number of words
}

