% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stats.R
\name{filter_by_choose}
\alias{filter_by_choose}
\title{filter by name, season, episode}
\usage{
filter_by_choose(all_scripts, the_name = "all", the_season = "all",
  the_episode = "all")
}
\arguments{
\item{all_scripts}{dataframe with all the scripts arranged}

\item{name}{a character vector. ("all") will mean no filter}

\item{season}{a numeric vector. ("all") will mean no filter}

\item{episode}{a numeric vector. ("all") will mean no filter}
}
\value{
dataframe filtered
}
\description{
A helper function to choose to filter by name/season/episode
}

