% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_list_from_rda_files}
\alias{get_list_from_rda_files}
\title{Merge all scripts saved in files to a list.}
\usage{
get_list_from_rda_files(thepath, nums)
}
\arguments{
\item{thepath}{The path where the Rda files are in}

\item{nums}{a numeric vector of the episodes numbers to read from the web.}
}
\value{
a list. A list where each item is a raw script.
}
\description{
This function merges all Rda files with the script to a list in case the connection to the web halts in the
middle of teh download
}

