% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Read all Seinfeld scripts to a list}
\usage{
get_raw_data(nums)
}
\arguments{
\item{nums}{a numeric vector of the episodes numbers to read from the web.}
}
\value{
a list. A list where each item is a raw script.
}
\description{
A function to read all the scripts from the web and save as files as well as generating a list
with the raw data
}

