% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_raw_episode}
\alias{get_raw_episode}
\title{Read a script of an episode from seinology.com}
\usage{
get_raw_episode(ep_num)
}
\arguments{
\item{ep_num}{a number. The number of an episode starting from the beginning of the show (from 1 to 180)}
}
\value{
a string. A long string with all the html code of the page.
}
\description{
A function that read a script from seinology.com by episode's number (total number)
}

