% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_text.R
\name{htmlToText}
\alias{htmlToText}
\title{Convert HTML to text}
\usage{
htmlToText(input, ...)
}
\description{
A function to convert html to text. \cr
}
\details{
Author: Tony Breyal \cr
Date: 2011-11-18 \cr
Modified: 2011-11-18
Description: Extracts all text from a webpage (aims to extract only the text you would see in a web browser) \cr
Packages Used: RCurl, XML \cr
Blog Reference: Not published \cr
Copyright (c) 2011, under the Creative Commons Attribution-NonCommercial 3.0 Unported (CC BY-NC 3.0) License \cr
For more information see: https://creativecommons.org/licenses/by-nc/3.0/ \cr
All rights reserved. \cr
}

