% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_the_list.R
\name{make_a_cleaned_list}
\alias{make_a_cleaned_list}
\title{Get a raw list and return a cleaned one using \code{\link{clean_the_episode}}.}
\usage{
make_a_cleaned_list(raw_data_list)
}
\arguments{
\item{raw_data_list}{a list with the raw data.}
}
\value{
a list with cleaned data
}
\description{
A function that takes a raw data list and returns a cleaned list.
}

