% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_clouds.R
\name{make_cloud_by_name}
\alias{make_cloud_by_name}
\title{Make a word cloud from a dataframe}
\usage{
make_cloud_by_name(all_scripts, name, stopwords = FALSE, season = "all",
  episode = "all")
}
\arguments{
\item{df}{a dataframe to make cloud from}
}
\value{
A word cloud chart
}
\description{
Make a word cloud from a dataframe
}

