% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_the_list.R
\name{merge_into_big_df}
\alias{merge_into_big_df}
\title{Merge an arranged list into one big dataframe}
\usage{
merge_into_big_df(arranged_list)
}
\arguments{
\item{arranged_list}{a list of arranged dataframes}
}
\value{
a dataframe with the arranged script of all episodes.
}
\description{
Merge an arranged list into one big dataframe
}

