% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stats.R
\name{plot_the_speakers}
\alias{plot_the_speakers}
\title{Plot a frequency plot}
\usage{
plot_the_speakers(freq, num)
}
\arguments{
\item{freq}{a frequency table}

\item{num}{a number indicates the number of bars in the plot}
}
\value{
a bar plot
}
\description{
Plot a frequency plot for the speakers
}

