% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helper_functions.R
\name{remove_after}
\alias{remove_after}
\title{Remove everything after a pattern}
\usage{
remove_after(string, pattern, occur = "last")
}
\arguments{
\item{string}{a string to remove pattern from}

\item{pattern}{a regular expression that indicates the pattern to remove}

\item{occur}{string indicates weather to remove from/till "first" or "last" occurance}
}
\value{
a string after removing everything after the pattern
}
\description{
Remove everything after a pattern in a string
}

