% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helper_functions.R
\name{remove_before}
\alias{remove_before}
\title{Remove everything before a pattern}
\usage{
remove_before(string, pattern, occur = "first")
}
\arguments{
\item{string}{a string to remove pattern from}

\item{pattern}{a regular expression that indicates the pattern to remove}

\item{occur}{string indicates weather to remove from/till "first" or "last" occurance}
}
\value{
a string after removing everything before the pattern
}
\description{
Remove everything before a pattern in a string
}

