% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helper_functions.R
\name{remove_between}
\alias{remove_between}
\title{Remove everything between two patterns}
\usage{
remove_between(string, pattern)
}
\arguments{
\item{string}{a string to remove pattern from}

\item{pattern}{a character matrix with 2 columns. The first element is the left pattern and the
second is the right (if the pattern needs escaping with backslash then add it.}
}
\value{
a string after removing everything between the two patterns
}
\description{
Remove everything between two patterns in a string
}

