% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helper_functions.R
\name{remove_exact}
\alias{remove_exact}
\title{Remove everything that matches a pattern}
\usage{
remove_exact(string, pattern)
}
\arguments{
\item{string}{a string to remove pattern from}

\item{pattern}{a regular expression that indicates the pattern to remove}
}
\value{
a string after removing everything that matches the pattern
}
\description{
Remove everything that matches a pattern in a string
}

