% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helper_functions.R
\name{remove_pattern}
\alias{remove_pattern}
\title{Remove a pattern}
\usage{
remove_pattern(string, pattern, type, occur)
}
\arguments{
\item{string}{a string to remove pattern from}

\item{pattern}{a regular expression that indicates the pattern to remove}

\item{type}{a string indicates wheater "before","after"}

\item{occur}{string indicates weather to remove from/till "first" or "last" occurance}
}
\value{
string after removing the pattern
}
\description{
Remove a pattern from a string with a type (before, after, between)
}

