% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helper_functions.R
\name{remove_punc}
\alias{remove_punc}
\title{Remove white spaces}
\usage{
remove_punc(string)
}
\arguments{
\item{string}{a string to remove spaces from}
}
\value{
a string after removing punctuation
}
\description{
Remove white spaces from a string
}

