% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helper_functions.R
\name{remove_white_spaces}
\alias{remove_white_spaces}
\title{Remove punctuation}
\usage{
remove_white_spaces(string)
}
\arguments{
\item{string}{a string to remove punctuation from}
}
\value{
a string after removing white spaces
}
\description{
Remove punctuation from a string
}

