% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helper_functions.R
\name{stem_words}
\alias{stem_words}
\title{Stem words using SnowballC}
\usage{
stem_words(string)
}
\arguments{
\item{string}{a character vector of words}
}
\value{
a character vector where each word is stemmed
}
\description{
Split a string into character vector of words
}

