% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stats.R
\name{count_the_speakers}
\alias{count_the_speakers}
\title{Count who speakes most number of times}
\usage{
count_the_speakers(all_scripts)
}
\arguments{
\item{all_scripts}{a dataframe with all the scripts data arranged}
}
\value{
a frequency table
}
\description{
A function that counts who speaks the most number of times
}

