% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helper_functions.R
\name{split_to_words}
\alias{split_to_words}
\title{Split a string into vector of words}
\usage{
split_to_words(string)
}
\arguments{
\item{string}{a string to split into words}
}
\value{
a character vector where each word is an item
}
\description{
Split a string into character vector of words
}

